# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  email?: string;
}

export interface GetCurrentAuthUserResponse {
  authenticated: boolean;
  user?: AuthUser;
}

export type BillerRecurrence =
  (typeof BillerRecurrence)[keyof typeof BillerRecurrence];

export const BillerRecurrence = {
  monthly: "monthly",
  biweekly: "biweekly",
  weekly: "weekly",
  "one-time": "one-time",
} as const;

export interface Biller {
  id: number;
  userId: string;
  name: string;
  category: string;
  websiteUrl?: string | null;
  typicalAmount?: number | null;
  dueDayOfMonth?: number | null;
  recurrence: BillerRecurrence;
  color?: string | null;
  icon?: string | null;
  createdAt: string;
}

export type CreateBillerBodyRecurrence =
  (typeof CreateBillerBodyRecurrence)[keyof typeof CreateBillerBodyRecurrence];

export const CreateBillerBodyRecurrence = {
  monthly: "monthly",
  biweekly: "biweekly",
  weekly: "weekly",
  "one-time": "one-time",
} as const;

export interface CreateBillerBody {
  name: string;
  category: string;
  websiteUrl?: string | null;
  typicalAmount?: number | null;
  dueDayOfMonth?: number | null;
  recurrence: CreateBillerBodyRecurrence;
  color?: string | null;
  icon?: string | null;
}

export type UpdateBillerBodyRecurrence =
  (typeof UpdateBillerBodyRecurrence)[keyof typeof UpdateBillerBodyRecurrence];

export const UpdateBillerBodyRecurrence = {
  monthly: "monthly",
  biweekly: "biweekly",
  weekly: "weekly",
  "one-time": "one-time",
} as const;

export interface UpdateBillerBody {
  name?: string;
  category?: string;
  websiteUrl?: string | null;
  typicalAmount?: number | null;
  dueDayOfMonth?: number | null;
  recurrence?: UpdateBillerBodyRecurrence;
  color?: string | null;
  icon?: string | null;
}

export type BillInstanceStatus =
  (typeof BillInstanceStatus)[keyof typeof BillInstanceStatus];

export const BillInstanceStatus = {
  unpaid: "unpaid",
  paid: "paid",
  scheduled: "scheduled",
  overdue: "overdue",
} as const;

export interface BillInstance {
  id: number;
  billerId: number;
  userId: string;
  amountDue: number;
  dueDate: string;
  status: BillInstanceStatus;
  confirmationNumber?: string | null;
  paidAt?: string | null;
  biller?: Biller;
  createdAt: string;
}

export type CreateBillInstanceBodyStatus =
  (typeof CreateBillInstanceBodyStatus)[keyof typeof CreateBillInstanceBodyStatus];

export const CreateBillInstanceBodyStatus = {
  unpaid: "unpaid",
  paid: "paid",
  scheduled: "scheduled",
  overdue: "overdue",
} as const;

export interface CreateBillInstanceBody {
  billerId: number;
  amountDue: number;
  dueDate: string;
  status: CreateBillInstanceBodyStatus;
  confirmationNumber?: string | null;
}

export type UpdateBillInstanceBodyStatus =
  (typeof UpdateBillInstanceBodyStatus)[keyof typeof UpdateBillInstanceBodyStatus];

export const UpdateBillInstanceBodyStatus = {
  unpaid: "unpaid",
  paid: "paid",
  scheduled: "scheduled",
  overdue: "overdue",
} as const;

export interface UpdateBillInstanceBody {
  amountDue?: number;
  dueDate?: string;
  status?: UpdateBillInstanceBodyStatus;
  confirmationNumber?: string | null;
  paidAt?: string | null;
}

export type IncomeEntryRecurrence =
  (typeof IncomeEntryRecurrence)[keyof typeof IncomeEntryRecurrence];

export const IncomeEntryRecurrence = {
  weekly: "weekly",
  biweekly: "biweekly",
  monthly: "monthly",
  "one-time": "one-time",
} as const;

export interface IncomeEntry {
  id: number;
  userId: string;
  label: string;
  amount: number;
  payDate: string;
  recurrence: IncomeEntryRecurrence;
  createdAt: string;
}

export type CreateIncomeEntryBodyRecurrence =
  (typeof CreateIncomeEntryBodyRecurrence)[keyof typeof CreateIncomeEntryBodyRecurrence];

export const CreateIncomeEntryBodyRecurrence = {
  weekly: "weekly",
  biweekly: "biweekly",
  monthly: "monthly",
  "one-time": "one-time",
} as const;

export interface CreateIncomeEntryBody {
  label: string;
  amount: number;
  payDate: string;
  recurrence: CreateIncomeEntryBodyRecurrence;
}

export interface BankAccount {
  id: number;
  userId: string;
  name: string;
  balance: number;
  institution?: string | null;
  updatedAt: string;
}

export interface CreateAccountBody {
  name: string;
  balance: number;
  institution?: string | null;
}

export interface DashboardSummary {
  totalBillsDue30Days: number;
  totalBalance: number;
  safetyGap: number;
  billsDueCount: number;
  overdueCount: number;
  upcomingBills: BillInstance[];
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  codeVerifier: string;
  state?: string;
  nonce?: string;
  redirectUri: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  sessionToken: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type GetBillInstancesParams = {
  month?: number;
  year?: number;
};