# lib/api-zod/src/generated/types/billInstance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Biller } from "./biller";
import type { BillInstanceStatus } from "./billInstanceStatus";

export interface BillInstance {
  id: number;
  billerId: number;
  userId: string;
  amountDue: number;
  dueDate: string;
  status: BillInstanceStatus;
  confirmationNumber?: string | null;
  paidAt?: string | null;
  biller?: Biller;
  createdAt: string;
}