# lib/api-zod/src/generated/types/createBillInstanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBillInstanceBodyStatus } from "./createBillInstanceBodyStatus";

export interface CreateBillInstanceBody {
  billerId: number;
  amountDue: number;
  dueDate: string;
  status: CreateBillInstanceBodyStatus;
  confirmationNumber?: string | null;
}