# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BillInstance } from "./billInstance";

export interface DashboardSummary {
  totalBillsDue30Days: number;
  totalBalance: number;
  safetyGap: number;
  billsDueCount: number;
  overdueCount: number;
  upcomingBills: BillInstance[];
}