# lib/api-zod/src/generated/types/incomeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
import type { IncomeEntryRecurrence } from "./incomeEntryRecurrence";

export interface IncomeEntry {
  id: number;
  userId: string;
  label: string;
  amount: number;
  payDate: string;
  recurrence: IncomeEntryRecurrence;
  createdAt: string;
}