# lib/api-zod/src/generated/types/logoutMobileSessionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */

export interface LogoutMobileSessionResponse {
  success: boolean;
}