# lib/api-zod/src/generated/types/updateBillerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrismClone Bill Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBillerBodyRecurrence } from "./updateBillerBodyRecurrence";

export interface UpdateBillerBody {
  name?: string;
  category?: string;
  websiteUrl?: string | null;
  typicalAmount?: number | null;
  dueDayOfMonth?: number | null;
  recurrence?: UpdateBillerBodyRecurrence;
  color?: string | null;
  icon?: string | null;
}